% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orden_app.R
\name{orden_app}
\alias{orden_app}
\title{Ordena los elementos de un json en base a la distancia de estos respecto a unas coordenadas de reerencia}
\usage{
orden_app(mi_posicion, elemento)
}
\arguments{
\item{mi_posicion, elemento}{}
}
\value{
json
}
\description{
Ordena los elementos de un json en base a la distancia de estos respecto a unas coordenadas de reerencia
}
\examples{
 orden_app("40.03695469525709,-6.081439474677546","Alojamientos")

}
